requirements.txt
# ========================================
# 🏥 医疗RAG套件依赖文件
# ========================================
# 注意：需要Python 3.9+ (LightRAG要求)
# 推荐安装命令：pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple

# ========================================
# 核心RAG框架
# ========================================
lightrag-hku>=0.1.0

# ========================================
# 模型服务
# ========================================
ollama

# ========================================
# 基础依赖
# ========================================
PyYAML>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# ========================================
# 文档处理
# ========================================
PyPDF2
python-docx
docx2txt
pandas>=2.0.0

# ========================================
# 文件监听
# ========================================
watchdog>=3.0.0

# ========================================
# 网络请求
# ========================================
httpx>=0.25.0
aiohttp
asyncio-throttle>=1.0.2

# ========================================
# 数据处理
# ========================================
numpy>=1.24.0

# ========================================
# 文本处理
# ========================================
jieba>=0.42.1
tiktoken>=0.5.0

# ========================================
# Web框架（可选API服务）
# ========================================
fastapi>=0.104.1
uvicorn>=0.24.0

# ========================================
# 工具库
# ========================================
tqdm>=4.66.0
aiofiles>=23.2.1

# ========================================
# 医疗文本处理（可选）
# ========================================
spacy>=3.7.0

# ========================================
# 图形和可视化（可选）
# ========================================
matplotlib>=3.7.2
networkx>=3.1

# ========================================
# 开发工具（可选）
# ========================================
black>=23.9.1
pytest>=7.4.3

# ========================================
# 其他依赖
# ========================================
pdfplumber>=0.9.0
openpyxl>=3.1.0

# ========================================
# 嵌入和向量数据库
# ========================================
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# ========================================
# 其他工具
# ========================================
pathlib
openai>=1.0.0

# ========================================
# 其他工具
# ========================================
requests>=2.31.0
typing-extensions>=4.8.0

# ========================================
# 其他工具
# ========================================
colorlog>=6.7.0
python-dateutil>=2.8.0
psutil>=5.9.0

# ========================================
# 其他工具
# ========================================
orjson>=3.9.0

# ========================================
# 其他工具
# ========================================
memory-profiler>=0.61.0 